feat: Initialize Hydro App project structure and logging system
- Added initial project files including README and notes.
- Implemented logging functionality using Boost.Log with custom formatting.
- Created a connection pool for PostgreSQL database management.
- Developed a query builder for ORM-like database interactions.
- Set up configuration management with JSON support.
- Enhanced console output with application details and version checks.
- Established a container for application settings and dependencies.
- Configured CMake for project build management.
- Updated VSCode settings for C/C++ development environment.

+++ build.sh
@@ -3,24 +3,26 @@
 LOGS_FOLDER="Logs"
 BUILD_FOLDER="build"
 SETTINGS_FILE="Settings.json"
+BUILD_LOG="build.log"
 
-if [ -d "$LOGS_FOLDER" ]; then
-    rm -rf "$LOGS_FOLDER"
-fi
+python3 tools/update/cmake.py
 
-if [ -d "$BUILD_FOLDER" ]; then
-    rm -rf "$BUILD_FOLDER"
-fi
-
-if [ -f "$SETTINGS_FILE" ]; then
-    rm -f "$SETTINGS_FILE"
-fi
+[[ -d "$LOGS_FOLDER" ]] && rm -rf "$LOGS_FOLDER"
+[[ -d "$BUILD_FOLDER" ]] && rm -rf "$BUILD_FOLDER"
+[[ -f "$SETTINGS_FILE" ]] && rm -f "$SETTINGS_FILE"
 
 mkdir "$BUILD_FOLDER"
 cd "$BUILD_FOLDER" || exit 1
 
-if cmake .. && make; then
-    clear
-else # if error occured does not clear the screen
+if cmake .. 2>&1 | tee "$BUILD_LOG"; then
+    if make 2>&1 | tee -a "$BUILD_LOG"; then
+        clear
+        mv -f compile_commands.json ..
+    else
+        echo "Make failed. See $BUILD_FOLDER/$BUILD_LOG for details."
+        exit 1
+    fi
+else
+    echo "CMake failed. See $BUILD_FOLDER/$BUILD_LOG for details."
     exit 1
 fi
